% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_govuk.R
\name{download_pages}
\alias{download_pages}
\title{Download the HTML source of pages published on gov.uk publications
\url{https://www.gov.uk/government/publications}}
\usage{
download_pages(directory)
}
\arguments{
\item{directory}{Specify directory for download}
}
\description{
Download the HTML source of pages published on gov.uk publications
\url{https://www.gov.uk/government/publications}
}
