% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{main_page}
\alias{main_page}
\title{Open main page of gov.UK publications
\url{https://www.gov.uk/government/publications}}
\usage{
main_page()
}
\description{
Open main page of gov.UK publications
\url{https://www.gov.uk/government/publications}
}
