% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{start_browser}
\alias{start_browser}
\title{Start Selenium driver}
\usage{
start_browser(port = 4567L, browser = c("chrome", "firefox", "phantomjs",
  "internet explorer"), verbose = TRUE)
}
\arguments{
\item{port}{Port to use}

\item{browser}{Which browser should be used
\describe{
\item{\code{"chrome"}}{}
\item{\code{"firefox"}}{}
\item{\code{"phantomjs"}}{}
\item{\code{"internet explorer"}}{}
}}

\item{verbose}{Print out messages}
}
\description{
\code{start_browser} should be run before execution of other functions in the
package. Selenium driver is required for both application of filters and 
parsing results. Further documentation is available in \code{\link[RSelenium]{rsDriver}}.
}
