% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{start_browser}
\alias{start_browser}
\title{Start Selenium driver}
\usage{
start_browser(port = 4445L, docker = TRUE, browser = c("chrome",
  "firefox"), verbose = TRUE, check = TRUE, extraCapabilities = list())
}
\arguments{
\item{port}{Port to use}

\item{docker}{Specify whether instead of launching its own server,
Docker container is runninng and can be connected to.}

\item{browser}{Which browser should be used
\describe{
\item{\code{"chrome"}}{}
\item{\code{"firefox"}}{}
}}

\item{verbose}{Print out messages}

\item{check}{Check the version of Selenium and associated drivers}

\item{extraCapabilities}{Optional arguments (browser-specific)}
}
\description{
\code{start_browser} should be run before execution of other functions in the
package. Selenium driver is required for both application of filters and 
parsing results. Further documentation is available in \code{\link[RSelenium]{rsDriver}}.
}
