% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{use_filter}
\alias{use_filter}
\alias{govuk_use_filter}
\title{Filter publications based on selection criteria}
\usage{
use_filter(selection, filter_type = c("menu", "keywords", "from_date",
  "to_date"))

govuk_use_filter(selection, filter_type = c("menu", "keywords", "from_date",
  "to_date"))
}
\arguments{
\item{selection}{Must be either a value that defines option from a drop-down
menu (retrieved with \link{get_filters}) or a text to be inserted in search
bar (keyword or date ranges)}

\item{filter_type}{Type of filter to which the selection is applied 
(drop-down menu by default)}
}
\description{
Filter publications based on selection criteria
}
\examples{
\dontrun{
use_filter("close-consultations")
use_filter("01/01/2018", filter_type = "from_date")
}

}
\author{
Tom Paskhalis
}
